Timer enable switches are 64-bit; Floppy reads/writes now correctly finish at gaps; Sector-based reads/writes no longer cause a delay when signaling finish.

+++ src/disc.c
@@ -37,7 +37,7 @@ int drive_empty[2] = {1, 1};
 int disc_changed[2];
 
 int motorspin;
-int motoron;
+int64_t motoron;
 
 int fdc_indexcount = 52;
